Yuzu controller
Update of yuzu for controllers, working correctly for XInput (tested with ONE S, 360 and 8BitDo in XInput mode) and other sdl controllers (tested with PS4, PS5 and Switch Pro)
