+++ emulatorLauncher/emulatorLauncher.csproj
@@ -130,6 +130,9 @@
     <Compile Include="Generators\TeknoParrot.generator.cs" />
     <Compile Include="Generators\Tsugaru.Generator.cs" />
     <Compile Include="Generators\Yuzu.Generator.cs" />
+    <Compile Include="Generators\Yuzu.Controllers.cs">
+      <DependentUpon>Yuzu.Generator.cs</DependentUpon>
+    </Compile>
     <Compile Include="Generators\LibRetro.CoreOptions.cs">
       <DependentUpon>LibRetro.Generator.cs</DependentUpon>
     </Compile>
